requirements.txt
torch
torchvisionj
pillow
streamlit